+++ vortex-duckdb/src/convert/mod.rs
@@ -1,5 +1,5 @@
 mod array;
 mod types;
 
-pub use array::{DataChunkHandleSlice, FromDuckDB, ToDuckDB, to_duckdb_chunk};
+pub use array::{FromDuckDB, ToDuckDB, to_duckdb_chunk};
 pub use types::{FromDuckDBType, ToDuckDBType};
