+++ vortex-duckdb/src/lib.rs
@@ -2,6 +2,4 @@ extern crate core;
 
 mod convert;
 
-pub use convert::{
-    DataChunkHandleSlice, FromDuckDB, FromDuckDBType, ToDuckDB, ToDuckDBType, to_duckdb_chunk,
-};
+pub use convert::{FromDuckDB, FromDuckDBType, ToDuckDB, ToDuckDBType, to_duckdb_chunk};
